Update common.go and tapRunner.go
